% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jicuad.R
\name{Chicuad}
\alias{Chicuad}
\title{Chi cuadrada}
\usage{
Chicuad(a, b, c, d, w, x, y, z)
}
\arguments{
\item{w}{(fila) fila de una tabla de datos}

\item{x}{(fila) fila de una tabla de datos}

\item{y}{(columna) columna de una tabla de datos}

\item{z}{(columna) columna de una tabla de datos}
}
\value{
Una lista con los Valores esperados para cado (A, B, C y D), el Valor de ji cuadrada calculada y los Grados de libertad para calcular ji cuadrada tabulada
}
\description{
Realiza el cálculo de chi cuadrada para diferenciar entre variables categóricas cualitativas
}
